// src/timesheet-web/src/app/api/models/backlog-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BacklogDto {
  assigneeUserId?: number | null;
  backlogCode?: string | null;
  createdAt?: string;
  deadlineExternal?: string | null;
  deadlineInternal?: string | null;
  endDate?: string | null;
  id?: number;
  note?: string | null;
  officialEstimateHours?: number | null;
  pcaContactId?: number | null;
  periodMonth?: string | null;
  progressPercent?: number | null;
  project?: string | null;
  roughEstimateHours?: number | null;
  rowVersion?: number;
  startDate?: string | null;
  teamId?: number | null;
  type?: string | null;
}